Prevent tail error on initial startup
The tail command throws a error message stating the log files do not exist when initially reading them, as it executes before they are created.  This resolves the issue by creating those log files first.  This did not affect functionality of the overall script, but the error message may cause users to believe there was an issue when everything was working fine.